Move validation module to 'modules' dir
