+++ src/app/app.module.ts
@@ -19,7 +19,7 @@ import { CountriesEffects } from './store/countries/countries.effects';
 import { NavigationButtonComponent } from './components/navigation-button/navigation-button.component';
 import * as fromCart from './store/cart/cart.reducer';
 import { CartEffects } from './store/cart/cart.effects';
-import { ValidationModule } from './shared/validation/validation.module';
+import { ValidationModule } from './modules/validation/validation.module';
 
 const appInitializerFactory =
   (appInitializer: AppInitializerService) => async (): Promise<void> => await appInitializer.initialize();
