+++ src/app/modules/registration/components/personal-information/personal-information.model.ts
@@ -1,5 +1,5 @@
 import { DynamicFormGroupModel, DynamicFormModel, DynamicInputModel } from '@ng-dynamic-forms/core';
-import  * as validationMessages from 'src/app/shared/validation/constants/validation-messages.constants';
+import  * as validationMessages from 'src/app/modules/validation/constants/validation-messages.constants';
 
 export const PERSONAL_INFORMATION_INPUTS = {
   firstName: new DynamicInputModel({
